# web/docs/swagger.yaml
basePath: /
definitions:
  errorinterface.ErrorResponse:
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  keycloak.PolicyRepresentation:
    properties:
      decisionStrategy:
        type: string
      description:
        type: string
      id:
        type: string
      logic:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  keycloak.ResourceRepresentation:
    properties:
      _id:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      displayName:
        type: string
      icon_uri:
        type: string
      name:
        type: string
      owner:
        additionalProperties:
          type: string
        type: object
      scopes:
        items:
          $ref: '#/definitions/keycloak.ScopeRepresentation'
        type: array
      type:
        type: string
      uris:
        items:
          type: string
        type: array
    type: object
  keycloak.ScopeRepresentation:
    properties:
      displayName:
        type: string
      iconUri:
        type: string
      id:
        type: string
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/keycloak.PolicyRepresentation'
        type: array
      resources:
        items:
          $ref: '#/definitions/keycloak.ResourceRepresentation'
        type: array
    type: object
  models.Action:
    properties:
      action:
        type: string
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
      description:
        type: string
    type: object
  models.AuditLog:
    properties:
      action:
        description: Action taken (e.g., "Created Task", "Updated Milestone")
        type: string
      details:
        description: Detailed description of the action
        type: string
      timestamp:
        description: When the action was performed
        type: string
      user_id:
        description: The user who performed the action
        type: string
    type: object
...
  models.Resource:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
      audit_logs:
        description: Audit logs for project actions
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        description: Flag to indicate if the project is deleted
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        description: Project owner (user ID)
        type: string
      project_id:
        type: string
      type:
        $ref: '#/definitions/models.ResourceType'
      updated_timestamp_utc:
        type: string
      urn:
        type: string
      version:
        type: string
    type: object
  models.ResourceType:
    enum:
    - db:mysql
    - app:github
    - llm:openai
    type: string
    x-enum-varnames:
    - ResourceTypeDatabase
    - ResourceTypeAPI
    - ResourceTypeLLM
...
  oauth.ResourceRequest:
    properties:
      _id:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      displayName:
        type: string
      icon_uri:
        type: string
      name:
        type: string
      owner:
        additionalProperties:
          type: string
        type: object
      scopes:
        items:
          $ref: '#/definitions/keycloak.ScopeRepresentation'
        type: array
      type:
        type: string
      uris:
        items:
          type: string
        type: array
    type: object
  oauth.ResourceResponse:
    properties:
      _id:
        type: string
      attributes:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      displayName:
        type: string
      icon_uri:
        type: string
      name:
        type: string
      owner:
        additionalProperties:
          type: string
        type: object
      scopes:
        items:
          $ref: '#/definitions/keycloak.ScopeRepresentation'
        type: array
      type:
        type: string
      uris:
        items:
          type: string
        type: array
    type: object
...
  resources.ResourceRequest:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
      audit_logs:
        description: Audit logs for project actions
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        description: Flag to indicate if the project is deleted
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        description: Project owner (user ID)
        type: string
      project_id:
        type: string
      type:
        $ref: '#/definitions/models.ResourceType'
      updated_timestamp_utc:
        type: string
      urn:
        type: string
      version:
        type: string
    type: object
  resources.ResourceResponse:
    properties:
      actions:
        items:
          $ref: '#/definitions/models.Action'
        type: array
      audit_logs:
        description: Audit logs for project actions
        items:
          $ref: '#/definitions/models.AuditLog'
        type: array
      created_timestamp_utc:
        type: string
      deleted:
        description: Flag to indicate if the project is deleted
        type: boolean
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        description: Project owner (user ID)
        type: string
      project_id:
        type: string
      type:
        $ref: '#/definitions/models.ResourceType'
      updated_timestamp_utc:
        type: string
      urn:
        type: string
      version:
        type: string
    type: object
  resources.ResourcesResponse:
    properties:
      resources:
        items:
          $ref: '#/definitions/models.Resource'
        type: array
    type: object
...
host: localhost:8002
info:
  contact: {}
paths:
...
  /{workspace_id}/{project_id}/oauth/resources:
    get:
      consumes:
      - application/json
      description: Lists all Keycloak resources with optional filtering
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by type
        in: query
        name: type
        type: string
      - description: Filter by URI
        in: query
        name: uri
        type: string
      - description: Filter by owner
        in: query
        name: owner
        type: string
      - description: Filter by scope
        in: query
        name: scope
        type: string
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/oauth.ResourceResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Creates a new Keycloak resource
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Resource Details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/oauth.ResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oauth.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Create a new resource
      tags:
      - resources
  /{workspace_id}/{project_id}/oauth/resources/{resource_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a Keycloak resource
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Delete a resource
      tags:
      - resources
    get:
      consumes:
      - application/json
      description: Retrieves a Keycloak resource by ID
      parameters:
      - description: Workspace ID
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/oauth.ResourceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      summary: Get a resource
      tags:
      - resources
...
  /projects/{project_id}/resources:
    get:
      consumes:
      - application/json
      description: Lists all resources for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.ResourcesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List resources by project
      tags:
      - resources
    post:
      consumes:
      - application/json
      description: Creates a new resource
      parameters:
      - description: Resource details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/resources.ResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create resource
      tags:
      - resources
  /projects/{project_id}/resources/{resource_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a resource
      parameters:
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete resource
      tags:
      - resources
    get:
      consumes:
      - application/json
      description: Gets a resource by ID
      parameters:
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get resource
      tags:
      - resources
    put:
      consumes:
      - application/json
      description: Updates an existing resource
      parameters:
      - description: Resource ID
        in: path
        name: resource_id
        required: true
        type: string
      - description: Updated resource details
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/resources.ResourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resources.ResourceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorinterface.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update resource
      tags:
      - resources
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"